Fixed #78
Attempted to update the post layout to accept the way we were defining
authors in the .md files, however it ended up being easier to modify the
way we were defining rather than change the layout.  All in all, I don't
think it will be a big issue to define them the way it is now.

We'll need to make sure we add an author to the _data/members.yml file for
each author that posts.  I've added an MHV author for generic posts such
as the quadcopter workshop, and infrastucture day posts.